requirements/base.txt
Django==1.7
dj-static==0.0.6
django-endless-pagination==2.0
django-gravatar2==1.1.3
Pillow==2.5.3
psycopg2==2.5.3
pytz==2014.4
git+http://github.com/nlhkabu/Django-parsley#egg=django-parsley


# Testing
factory-boy==2.4.1 # Mocks
PyYAML==3.11 # Fixtures
selenium
behave==1.2.4
git+http://github.com/rwillmer/django-behave#egg=django-behave
parse
splinter

# Deployment
uWSGI==2.0.7
